Added many more features and 4x the code length.
Added the last input feature without that it would only be 2x. For the sake of not dealing with stupid shit since the code will be the same other than the exits and indentation i have decided to copy and paste the exit code and put it into a different text doccument on my computer for IDE opening. Please do the same future devs.